+++ standalone/src/main/java/software/xdev/mockserver/mock/RequestMatchers.java
@@ -35,7 +35,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static software.xdev.mockserver.log.model.LogEntryMessages.*;
+import static software.xdev.mockserver.logging.LoggingMessages.*;
 import static software.xdev.mockserver.util.StringUtils.isBlank;
 import static software.xdev.mockserver.util.StringUtils.isNotBlank;
 import static software.xdev.mockserver.mock.SortableExpectationId.EXPECTATION_SORTABLE_PRIORITY_COMPARATOR;